Clean up deprecated code from refactor cycle
Deletes the roster.go and roster_test.go files, deprecated during the
refactor. Likewise, the lines for helper functions in
roster_view_test.go (now contained in the testkit).

Updates the Sequence type on RecordedRosterEvent to int64.

+++ internal/application/events.go
@@ -9,6 +9,6 @@ import (
 var ErrUnrecognizedRecordedEvent = errors.New("unrecognized recorded event")
 
 type RecordedRosterEvent struct {
-	Sequence int
+	Sequence int64
 	Event    domain.RosterEvent
 }
